updated feature engineering

+++ src/models/train_model.py
@@ -8,7 +8,7 @@ from sklearn.ensemble import GradientBoostingClassifier
 train_data = pd.read_csv('./data/features/train_tfidf.csv')
 
 X_train = train_data.iloc[:,0:-1].values
-y_train = train_data.iloc[:,-1].values
+y_train = train_data.iloc[:,-1].values.astype(int)
 
 # Define and train the XGBoost model
